pyproject.toml
[project]
name = "ds-masters-2025-llm-rooting"
version = "0.1.0"
description = "Data Science Masters Project"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "uvicorn==0.23.2",
    "langchain-openai>=0.2.1",
    "kaleido>=0.2.1",
    "plotly>=5.24.1",
    "numpy>=1.26.4",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.0",
    "google-cloud-bigquery>=3.26.0",
    "toml>=0.10.2",
    "langchain-google-vertexai>=2.0.26",
    "google-cloud-datastore>=2.20.1",
    "fastapi>=0.115.4",
    "pandas>=2.2.3",
    "langchain-google-community>=2.0.2",
    "pydantic>=2.9.1",
    "pytest-mock>=3.14.0",
    "scikit-learn>=1.5.2",
    "db-dtypes>=1.3.1",
    "mlflow==3.1.1",
    "langchain-ollama>=0.2.3",
    "litellm>=1.59.12",
    "langchain-qdrant>=0.2.0",
    "typer>=0.15.1",
    "anthropic[vertex]>=0.45.2",
    "langchain>=0.3.7",
    "langgraph>=0.4.8",
    "langchain-huggingface>=0.1.2",
    "google-adk>=0.0.1",
    "psycopg2>=2.9.10",
    "langchain-community>=0.3.26",
    "mistralai>=1.8.2",
    "langchain-mistralai>=0.2.10",
    "langchain-litellm>=0.2.1",
    "isort>=6.0.1",
    "nbformat>=5.10.4",
    "transformers>=4.53.3",
    # Remove these CPU-only PyTorch entries:
    # "torch>=2.0.0",
    # "torchvision>=0.15.0",
    # "torchvision>=0.15.0",
    # "torchaudio>=2.0.0",
    "fastmcp>=2.10.6",
    "langchain-mcp-adapters>=0.1.9",
    "mcp>=1.12.2",
    "seaborn>=0.13.2",
    "selenium>=4.35.0",
    "webdriver-manager>=4.0.2",
]

[project.optional-dependencies]
dev = [
    "isort>=5.13.2",
    "black>=24.10.0",
]
test = [
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.isort]
profile = "black"
src_paths = ["src"]
known_first_party = ["src"]
